Feat : 댓글,게시물 좋아요 여부, 게시물 스크랩 여부 확인 응답

+++ server/src/main/java/com/planmate/server/dto/response/post/PostPageResponseDto.java
@@ -10,10 +10,10 @@ import java.util.List;
 @AllArgsConstructor
 @Builder
 public class PostPageResponseDto {
-    private Long totalPages;
+    private Integer totalPages;
     List<PostResponseDto> postDtoList;
 
-    public static PostPageResponseDto of(Long totalPages,List<PostResponseDto> postDtoList) {
+    public static PostPageResponseDto of(Integer totalPages,List<PostResponseDto> postDtoList) {
         return PostPageResponseDto.builder()
                 .totalPages(totalPages)
                 .postDtoList(postDtoList)
